Refactor: [FE] api boundary 위치 변경
